Add FSR1 spatial upscaler & resolution scaling
Implement optional post-FX resolution scaling with FSR1 (EASU+RCAS) upsampling or bilinear fallback. Scene pass renders at configurable scale (default 0.67), post-FX bakes to same scale, then FSR1 upscales to native resolution after FXAA. Added dev panel controls for method, sharpness, and denoise. Downgrade TypeScript to 6.0.3 (from RC). Fix TSL type casts in grass SSBO packing and terrain shading (uint() wraps for bit operations).

+++ src/world/grass/compute/grassSsboPack.ts
@@ -54,7 +54,7 @@ export function packStateWord(
   const vis = uint(visByte);
   const sc = encodeScale12(currentScale, scaleMin, scaleSpan);
   const so = encodeScale12(originalScale, scaleMin, scaleSpan);
-  return vis.add(shiftLeft(sc, 8)).add(shiftLeft(so, 20));
+  return vis.add(shiftLeft(uint(sc), 8)).add(shiftLeft(uint(so), 20));
 }
 
 export function encodeVisBool(visibility: TslNode): TslNode {
@@ -66,11 +66,11 @@ export function unpackVisByte(word: TslNode): TslNode {
 }
 
 export function unpackCurrentScale(word: TslNode, scaleMin: TslNode, scaleSpan: TslNode): TslNode {
-  return decodeScale12(bitAnd(shiftRight(word, 8), MASK12), scaleMin, scaleSpan);
+  return decodeScale12(bitAnd(uint(shiftRight(word, 8)), MASK12), scaleMin, scaleSpan);
 }
 
 export function unpackOriginalScale(word: TslNode, scaleMin: TslNode, scaleSpan: TslNode): TslNode {
-  return decodeScale12(bitAnd(shiftRight(word, 20), MASK12), scaleMin, scaleSpan);
+  return decodeScale12(bitAnd(uint(shiftRight(word, 20)), MASK12), scaleMin, scaleSpan);
 }
 
 function encodeHeight16(heightNorm: TslNode): TslNode {
